Print and plot oolong summary

These functions print or plot a useful summary of the results from summarize_oolong. For details, please see the overview vignette: vignette("overview", package = "oolong")

# S3 method for oolong_summary
print(x, ...)

# S3 method for oolong_summary
plot(x, ...)

Arguments

x

an oolong_summary

...

other parameters

Value

None

Diagnostic plot

Plot function displays a diagnostic plot with the following subplots (gs only).

Top left

Correlation between answer from coders and target value to check for correlation between two values. Both axes are minmax transformed.

Top right

Bland-altman plot of answer from coders and target value to check for agreement between two values.

Bottom left

Correlation between target value and content length to check for the influence of content length.

Bottom right

Cook's distance to check for influential observations.

Author

Chung-hong Chan